fix: moved paymentMethod and Pricing pages components folder. Modal component and used methods refactored.
