+++ hooks/useStore.ts
@@ -4,8 +4,15 @@ import { persist, createJSONStorage } from 'zustand/middleware'
 import { ModalState, UserState } from '@/types/store.types'
 
 export const useModalStore = create<ModalState>((set) => ({
-  showModal: false,
-  toggleModal: () => set((state) => ({ showModal: !state.showModal }))
+  showModal: {},
+  toggleModal: (modalName) =>
+    set((state) => ({
+      showModal: {
+        ...state.showModal,
+        [modalName]: !state.showModal[modalName] // toggleModal(key) triggerlandığında gönderilen key'e ait modalı açmak veya kapatma işlemi
+      }
+    })),
+  closeAllModals: () => set({ showModal: {} }) // Tüm modalları kapatma işlemi
 }))
 
 export const useUserStore = create<UserState>()(
