+++ types/store.types.ts
@@ -1,6 +1,7 @@
 export type ModalState = {
-  showModal: boolean
-  toggleModal: () => void
+  showModal: Record<string, boolean> // Parametre adına göre boolean değerleri içerecek bir obje
+  toggleModal: (modalName: string) => void // Parametre adına göre modalı açıp kapatmak için
+  closeAllModals: () => void // Tüm modalları kapatmak için
 }
 
 export type UserState = {
